src/requirements.txt
actionlib==1.14.0
angles==1.9.13
appdirs==1.4.3
apturl==0.5.2
asttokens==2.4.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
bcrypt==3.1.7
black==19.3b0
blinker==1.4
bondpy==1.8.6
Brlapi==0.7.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.8.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
executing==2.1.0
filelock==3.9.0
fsspec==2023.4.0
future==0.18.2
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==8.5.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.29.5
ipython==8.12.3
isort==4.3.4
jedi==0.19.1
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jupyter-client==8.6.3
jupyter-core==5.7.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.3
matplotlib==3.1.2
matplotlib-inline==0.1.7
message-filters==1.17.0
more-itertools==4.2.0
mpi4py==4.0.0
mpmath==1.3.0
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.0
nose==1.3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
packaging==24.1
pandas==2.0.3
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
protobuf==3.6.1
psutil==5.5.1
pure-eval==0.2.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.8
roman==2.0.0
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.25
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.6
stack-data==0.6.3
sympy==1.12
systemd-python==234
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
toml==0.10.0
topic-tools==1.17.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4.1
traitlets==5.14.3
triton==2.1.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.4.0
tzdata==2024.2
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.13
wsaccel==0.6.2
xacro==1.14.19
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
